Adapted old changes of firestation for mirror calibration to current
master branch.

Conflicts:
	res/guis/firestation/firestation.ui
	src/libs/Makefile
	src/libs/fvmodels/Makefile
	src/tools/vision/firestation/firestation.cpp

+++ src/libs/fvcams/bumblebee2.cpp
@@ -34,6 +34,8 @@
 #include <string>
 #ifdef __FreeBSD__
 #  include <sys/endian.h>
+#elif defined(__MACH__) && defined(__APPLE__)
+#  include <sys/_endian.h>
 #else
 #  include <endian.h>
 #endif
@@ -231,7 +233,7 @@ Bumblebee2Camera::print_info()
 
   printf("Serial: %u\n", serial_no());
 #if __WORDSIZE == 64
-  printf("GUID:   0x%016lx\n", guid());
+  printf("GUID:   0x%016lx\n", (long unsigned int)guid());
 #else
   printf("GUID:   0x%016llx\n", guid());
 #endif
